Fix race conditions with asynchronous cache stores
Because TaxCloudCalculator#tax_for_item wrote to and then immediately
read from the cache to compute its return value, certain asynchronous
cache stores (e.g. Memcached) would fail.

More in general, this is a flawed approach, because we can't know
what happens between the cache write and the subsequent read.

Instead, we are now writing to the cache but returning the value
from memory.